Email capability
Note that you need to add your own email username & password for it to
work. also remember to change the "TO" field to send to yourself so u
can see it working. currently it sends an email 1 minute after it is
triggered. forgot when it is triggered. either on a button press or on
page load. did it this way jsut for testing